+++ src/main/java/com/example/whatseoul/controller/citydata/CityDataViewController.java
@@ -27,7 +27,6 @@ public class CityDataViewController {
 
 	@GetMapping("/")
 	public String indexPage(@AuthenticationPrincipal CustomUserDetails userDetails, Model model) {
-
 		if (userDetails != null) {
 			model.addAttribute("isAuthenticated", true);
 		} else {
